+++ examples/svelte-ts/src/routes/avatar.svelte
@@ -2,7 +2,7 @@
   import * as avatar from "@zag-js/avatar"
   import { avatarData } from "@zag-js/shared"
   import { normalizeProps, useMachine } from "@zag-js/svelte"
-  import Toolbar from "$lib/components/_toolbar.svelte"
+  import Toolbar from "$lib/components/toolbar.svelte"
 
   const images = avatarData.full
   const getRandomImage = () => images[Math.floor(Math.random() * images.length)]
