+++ examples/svelte-ts/src/routes/carousel.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Toolbar from "$lib/components/_toolbar.svelte"
+  import Toolbar from "$lib/components/toolbar.svelte"
   import { useControls } from "$lib/use-controls.svelte"
   import * as carousel from "@zag-js/carousel"
   import { carouselControls, carouselData } from "@zag-js/shared"
