+++ examples/svelte-ts/src/routes/index.svelte
